Adding hooks to the CRL to support the new BMOD software.
This will insert new BMW words into the datastream for the new values
available from the BMW cfSocket connection.
It will turn on FlexIO bit 4 (0x8) when the crate sees the FFB pause signal.
It will turn on FlexIO bit 3 (0x4) simultaniious with the modulation trigger,
and turn it off after the specified number of modulation periods.