Modified the CH/RHRS/LHRS parity CRL so that the VQWK modules are read the same as in the injector
The VQWK readout will be read at each event, even if they do not have new data.  The data from the
previous event will be resent; this is detected by the analyzer as an error condition.  This is how
the modules in the injector have been treated for the past several years.